requirements.txt
langchain==0.3.26
langchain-community==0.3.27
langchain-experimental==0.3.4
langchain-core==0.3.69
langchain-openai==0.3.28
python-dotenv==1.1.1
fastapi==0.116.1
uvicorn==0.35.0
pandas==2.3.1
tabulate==0.9.0
matplotlib==3.10.3
seaborn==0.13.2
python-multipart==0.0.20
streamlit==1.47.1
requests==2.32.4
openpyxl==3.1.5